콜로니 프로젝트 - 진행 작업 표시 구현

+++ src/main/java/org/duckdns/hjow/samples/colonyman/elements/facilities/Arcade.java
@@ -123,7 +123,7 @@ public class Arcade extends DefaultFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(240);
     }
     
     public static Long getTechNeeded() {
