+++ src/main/java/org/duckdns/hjow/samples/colonyman/elements/facilities/PowerStation.java
@@ -114,7 +114,7 @@ public class PowerStation extends DefaultFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1800);
+        return new Integer(300);
     }
     
     public static Long getTechNeeded() {
