+++ src/main/java/org/duckdns/hjow/samples/colonyman/elements/facilities/ResearchCenter.java
@@ -97,7 +97,7 @@ public class ResearchCenter extends DefaultFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1800);
+        return new Integer(1200);
     }
     
     public static Long getTechNeeded() {
