+++ src/main/java/org/duckdns/hjow/samples/colonyman/elements/facilities/Restaurant.java
@@ -148,7 +148,7 @@ public class Restaurant extends DefaultFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(180);
     }
     
     public static Long getTechNeeded() {
